intentando redirigir a log in si no estas logged

+++ src/main/java/com/example/demo/controller/OfferController.java
@@ -72,7 +72,12 @@ public class OfferController {
     }
 
     @RequestMapping(value="/update/{id_O}", method = RequestMethod.GET)
-    public String editOffer(Model model, @PathVariable int id_O) {
+    public String editOffer(Model model, @PathVariable int id_O, HttpSession session) {
+        if (session.getAttribute("student") == null) {
+            model.addAttribute("student", new Student());
+            session.setAttribute("nextUrl", "/offer/update");
+            return "login";
+        }
         model.addAttribute("offer", offerDao.getOffer(id_O));
         return "offer/update";
     }
@@ -80,15 +85,23 @@ public class OfferController {
     @RequestMapping(value="/update", method = RequestMethod.POST)
     public String processUpdateSubmit(
             @ModelAttribute("offer") Offer offer,
-            BindingResult bindingResult) {
+            BindingResult bindingResult, HttpSession session){
         if (bindingResult.hasErrors())
             return "offer/update";
+        if (session.getAttribute("student") == null) {
+            session.setAttribute("nextUrl", "/offer/delete");
+            return "login";
+        }
         offerDao.updateOffer(offer);
         return "redirect:list";
     }
 
     @RequestMapping(value="/delete/{id_O}")
-    public String processDelete(@PathVariable int id_O) {
+    public String processDelete(@PathVariable int id_O, HttpSession session) {
+        if (session.getAttribute("student") == null) {
+            session.setAttribute("nextUrl", "/offer/delete");
+            return "login";
+        }
         offerDao.deleteOffer(id_O);
         return "redirect:../list";
     }
